PAAC-479: Renamed classes and tests. Re-enabled PeriodCalculatorFactorySpec
